refactor: extract possession and pressure stats accumulators
