+++ src/stats/calculators/mod.rs
@@ -139,6 +139,8 @@ pub mod player_vertical_state;
 pub use player_vertical_state::*;
 pub mod possession;
 pub use possession::*;
+pub mod possession_stats;
+pub use possession_stats::*;
 pub mod possession_state;
 pub use possession_state::*;
 pub mod powerslide;
@@ -147,6 +149,8 @@ pub mod powerslide_stats;
 pub use powerslide_stats::*;
 pub mod pressure;
 pub use pressure::*;
+pub mod pressure_stats;
+pub use pressure_stats::*;
 pub mod rotation;
 pub use rotation::*;
 pub mod rotation_stats;
