+++ src/stats/calculators/possession.rs
@@ -19,10 +19,6 @@ impl PossessionStateLabel {
             Self::Neutral => "neutral",
         }
     }
-
-    fn as_label(self) -> StatLabel {
-        StatLabel::new("possession_state", self.as_label_value())
-    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -52,88 +48,6 @@ impl FieldThirdLabel {
             Self::TeamOneThird => "team_one_third",
         }
     }
-
-    fn as_label(self) -> StatLabel {
-        StatLabel::new("field_third", self.as_label_value())
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize)]
-pub struct PossessionStats {
-    pub tracked_time: f32,
-    pub team_zero_time: f32,
-    pub team_one_time: f32,
-    pub neutral_time: f32,
-    #[serde(default, skip_serializing_if = "LabeledFloatSums::is_empty")]
-    pub labeled_time: LabeledFloatSums,
-}
-
-impl PossessionStats {
-    pub fn team_zero_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.team_zero_time * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn team_one_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.team_one_time * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn neutral_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.neutral_time * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn time_with_labels(&self, labels: &[StatLabel]) -> f32 {
-        self.labeled_time.sum_matching(labels)
-    }
-
-    pub fn for_team(&self, is_team_zero: bool) -> PossessionTeamStats {
-        let (possession_time, opponent_possession_time) = if is_team_zero {
-            (self.team_zero_time, self.team_one_time)
-        } else {
-            (self.team_one_time, self.team_zero_time)
-        };
-
-        let mut labeled_time = LabeledFloatSums::default();
-        for entry in &self.labeled_time.entries {
-            labeled_time.add(
-                entry
-                    .labels
-                    .iter()
-                    .map(|label| team_relative_possession_label(label, is_team_zero)),
-                entry.value,
-            );
-        }
-
-        PossessionTeamStats {
-            tracked_time: self.tracked_time,
-            possession_time,
-            opponent_possession_time,
-            neutral_time: self.neutral_time,
-            labeled_time,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize, ts_rs::TS)]
-#[ts(export)]
-pub struct PossessionTeamStats {
-    pub tracked_time: f32,
-    pub possession_time: f32,
-    pub opponent_possession_time: f32,
-    pub neutral_time: f32,
-    #[serde(default, skip_serializing_if = "LabeledFloatSums::is_empty")]
-    pub labeled_time: LabeledFloatSums,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, ts_rs::TS)]
@@ -142,38 +56,9 @@ pub struct PossessionEvent {
     pub time: f32,
     pub frame: usize,
     pub active: bool,
+    pub duration: f32,
     pub possession_state: String,
     pub field_third: Option<String>,
-}
-
-fn team_relative_possession_label(label: &StatLabel, is_team_zero: bool) -> StatLabel {
-    match (label.key, label.value) {
-        ("possession_state", "team_zero") => StatLabel::new(
-            "possession_state",
-            if is_team_zero { "own" } else { "opponent" },
-        ),
-        ("possession_state", "team_one") => StatLabel::new(
-            "possession_state",
-            if is_team_zero { "opponent" } else { "own" },
-        ),
-        ("field_third", "team_zero_third") => StatLabel::new(
-            "field_third",
-            if is_team_zero {
-                "defensive_third"
-            } else {
-                "offensive_third"
-            },
-        ),
-        ("field_third", "team_one_third") => StatLabel::new(
-            "field_third",
-            if is_team_zero {
-                "offensive_third"
-            } else {
-                "defensive_third"
-            },
-        ),
-        _ => label.clone(),
-    }
 }
 
 #[derive(Debug, Clone, Default, PartialEq)]
@@ -337,7 +222,7 @@ impl PossessionTracker {
 
 #[derive(Debug, Clone, Default, PartialEq)]
 pub struct PossessionCalculator {
-    stats: PossessionStats,
+    stats: PossessionStatsAccumulator,
     tracker: PossessionTracker,
     events: EventStream<PossessionEvent>,
     last_emitted_event_state: Option<PossessionEventState>,
@@ -356,7 +241,7 @@ impl PossessionCalculator {
     }
 
     pub fn stats(&self) -> &PossessionStats {
-        &self.stats
+        self.stats.stats()
     }
 
     pub fn events(&self) -> &[PossessionEvent] {
@@ -367,30 +252,11 @@ impl PossessionCalculator {
         self.events.new_events()
     }
 
-    fn apply_possession_time(
-        stats: &mut PossessionStats,
-        state: PossessionStateLabel,
-        field_third: Option<FieldThirdLabel>,
-        dt: f32,
-    ) {
-        match state {
-            PossessionStateLabel::TeamZero => stats.team_zero_time += dt,
-            PossessionStateLabel::TeamOne => stats.team_one_time += dt,
-            PossessionStateLabel::Neutral => stats.neutral_time += dt,
-        }
-        if let Some(field_third) = field_third {
-            stats
-                .labeled_time
-                .add([state.as_label(), field_third.as_label()], dt);
-        } else {
-            stats.labeled_time.add([state.as_label()], dt);
-        }
-    }
-
     fn emit_event_if_changed(
         &mut self,
         frame: &FrameInfo,
         active: bool,
+        duration: f32,
         possession_state: PossessionStateLabel,
         field_third: Option<FieldThirdLabel>,
     ) {
@@ -399,16 +265,19 @@ impl PossessionCalculator {
             possession_state,
             field_third,
         };
-        if self.last_emitted_event_state == Some(event_state) {
+        if self.last_emitted_event_state == Some(event_state) && duration == 0.0 {
             return;
         }
-        self.events.push(PossessionEvent {
+        let event = PossessionEvent {
             time: frame.time,
             frame: frame.frame_number,
             active,
+            duration,
             possession_state: possession_state.as_label_value().to_owned(),
             field_third: field_third.map(|label| label.as_label_value().to_owned()),
-        });
+        };
+        self.stats.apply_event(&event);
+        self.events.push(event);
         self.last_emitted_event_state = Some(event_state);
     }
 
@@ -421,11 +290,10 @@ impl PossessionCalculator {
     ) -> SubtrActorResult<()> {
         self.events.begin_update();
         if !live_play_state.is_live_play {
-            self.emit_event_if_changed(frame, false, PossessionStateLabel::Neutral, None);
+            self.emit_event_if_changed(frame, false, 0.0, PossessionStateLabel::Neutral, None);
             return Ok(());
         }
 
-        self.stats.tracked_time += frame.dt;
         let field_third = ball.sample().map(FieldThirdLabel::from_ball);
         let state =
             if let Some(possession_team_is_team_0) = possession_state.active_team_before_sample {
@@ -437,8 +305,7 @@ impl PossessionCalculator {
             } else {
                 PossessionStateLabel::Neutral
             };
-        Self::apply_possession_time(&mut self.stats, state, field_third, frame.dt);
-        self.emit_event_if_changed(frame, true, state, field_third);
+        self.emit_event_if_changed(frame, true, frame.dt, state, field_third);
         Ok(())
     }
 }
